damage popups, some audio

+++ game_survivors/src/modules/core_animations/wiggle/components.hpp
@@ -7,7 +7,7 @@ namespace game2d {
 
 struct WiggleUpAndDown
 {
-  glm::ivec2 base_position;
+  glm::vec2 base_position;
   float time = 0.0f;
   float offset = 0.0f;
